A small tech demo of basic projection.
I was reminded in this year's Global Game Jam just how nice Lua Love is; I was
also reminded that it's really for 2d games.

I feel a side-scroller wouldn't provide enough room for the multiple players
and level of action I'd like, and top-down would lose some of the graphical
coolness.  True 3d systems are way out of my abilities, though, so I spent some
time looking around at examples of various "2.5d" systems.

I think the offset angle of isomorphic games would work well, but a straight-on
viewpoint is simpler, and thus first.

One of the things we had an issue with in Heart Healer Extreme was a statue (or
fountain, or something) in the center of the arena - if we allowed the
collision detection to work based on feet, rather than a flattening of
everything, our character would end up walking *on top* of the obstruction,
rather than behind it.  So, I thought about it a little and decided to play
around with solving that problem; that solution is this.  It's very simple, and
not at all ideal, but it works (although looks a bit funky because the lack of
collision detection allows blocks to move through each other).

'w' and 's' move the blocks up and down, and 1, 2, 3 select the block to move.